backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1
neo4j==5.20.0
google-cloud-storage==2.17.0
google-generativeai==0.5.4
httpx==0.27.0
python-dotenv==1.0.1
tenacity==8.2.3
structlog==24.1.0
